requirements.txt
# Core dependencies
prometheus-api-client==0.5.3
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
loguru==0.7.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Type checking
pydantic>=2.7.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# LLM Integration
groq==0.4.1